Fix #14 : 토큰 거래 데이터 재구성에 따른 DTO 재구성
