+++ src/main/java/com/ddiring/Backend_BlockchainConnector/domain/dto/trade/BuyInfo.java
@@ -1,0 +1,13 @@
+package com.ddiring.Backend_BlockchainConnector.domain.dto.trade;
+
+import lombok.*;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class BuyInfo {
+    private Long buyId;
+    private String buyerAddress;
+    private Long tokenAmount;
+}
