+++ src/main/java/com/ddiring/Backend_BlockchainConnector/domain/dto/trade/SellInfo.java
@@ -1,0 +1,13 @@
+package com.ddiring.Backend_BlockchainConnector.domain.dto.trade;
+
+import lombok.*;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class SellInfo {
+    private Long sellId;
+    private String sellerAddress;
+    private Long tokenAmount;
+}
