+++ src/main/java/com/ddiring/Backend_BlockchainConnector/domain/dto/trade/TradeDto.java
@@ -1,0 +1,25 @@
+package com.ddiring.Backend_BlockchainConnector.domain.dto.trade;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import lombok.*;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class TradeDto {
+    @NotNull(message = "거래 번호는 필수입니다.")
+    @Positive(message = "거래 번호는 양수여야 합니다.")
+    private Long tradeId;
+
+    @NotBlank(message = "사용자 주소는 필수입니다.")
+    private String smartContractAddress;
+
+    @NotNull
+    private BuyInfo buyInfo;
+
+    @NotNull
+    private SellInfo sellInfo;
+}
